+++ src/test/java/com/hermes/lotcenter/TestLotteryBetClient.java
@@ -30,7 +30,7 @@ public class TestLotteryBetClient {
 
         String lotCode = "FFK3";
         String qiHao = "202209071316";
-        Integer money = 1;
+        Double money = 1D;
         String name = "双";
         BetDataIdEnum dataIdEnum = BetDataIdEnum.fromName(name);
         Map<String, ?> criteria = doBetDomain.buildFromCriteria(lotCode, qiHao, money, dataIdEnum);
